# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DR TODAY API Documentation",
        "description": "API documentation for DR TODAY application",
        "contact": {
            "email": "admin@dr-today.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "DR TODAY API Server"
        }
    ],
    "paths": {
...
        "/api/driver_documents": {
            "get": {
                "tags": [
                    "Driver_document"
                ],
                "summary": "Get all Driver_documents",
                "operationId": "e14be14e8190c8a66b037ee73c7f1d70",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "post": {
                "tags": [
                    "Driver_document"
                ],
                "summary": "Create a new Driver_documents",
                "operationId": "7f84bd975cb8faa2fd2416c7146ce473",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/driver_documents/{driver_document}": {
            "get": {
                "tags": [
                    "Driver_document"
                ],
                "summary": "Get all {driver_document}",
                "operationId": "ebba6056194eec7a76c26a467b650b2e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "put": {
                "tags": [
                    "Driver_document"
                ],
                "summary": "Update a {driver_document} by ID",
                "operationId": "80362bfa9c47d9636128df8bd39104fa",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/driver_documents/delete/{id}": {
            "post": {
                "tags": [
                    "Driver_document"
                ],
                "summary": "Create a new Delete",
                "operationId": "d690d51ef66c9fe91cd5cee6d94093db",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/driver_documents/restore/{id}": {
            "post": {
                "tags": [
                    "Driver_document"
                ],
                "summary": "Create a new Restore",
                "operationId": "0ff1a7edae945382ec5799ec3eb0431e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}